% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conSV.r
\name{format_brass}
\alias{format_brass}
\title{Title: Format Brass}
\usage{
format_brass(brass)
}
\arguments{
\item{brass}{Data frame with BRASS output}
}
\description{
This function extract required fields from output vcf
}
