% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conSV.r
\name{format_caller}
\alias{format_caller}
\title{Title: Format output in bedpe}
\usage{
format_caller(df)
}
\arguments{
\item{df}{from formatted function}
}
\description{
This function format each caller output in standars bedpe
}
